+++ src/media/writer.rs
@@ -1,4 +1,3 @@
-use std::sync::Arc;
 use std::time::Instant;
 
 use crate::RtcError;
@@ -9,6 +8,7 @@ use crate::rtp_::VideoOrientation;
 use crate::session::Session;
 
 use super::{ExtensionValues, KeyframeRequestKind, Media, MediaTime, Mid, Pt, Rid, ToPayload};
+use crate::SharedBytes;
 
 /// Writer of frame level data.
 ///
@@ -139,7 +139,7 @@ impl<'a> Writer<'a> {
         pt: Pt,
         wallclock: Instant,
         rtp_time: MediaTime,
-        data: impl Into<Arc<[u8]>>,
+        data: impl Into<SharedBytes>,
     ) -> Result<(), RtcError> {
         // This (indirect) unwrap is OK due to the invariant of self.mid being resolvable
         let media = media_by_mid_mut(&mut self.session.medias, self.mid);
@@ -154,7 +154,7 @@ impl<'a> Writer<'a> {
             }
         }
 
-        let data: Arc<[u8]> = data.into();
+        let data: SharedBytes = data.into();
 
         trace!(
             "write {:?} {:?} {:?} time: {:?} len: {}",